update dependencies
Simplified some dependency use

minor simplify

update gitignore
