+++ .gitignore
@@ -404,3 +404,4 @@ riderModule.iml
 .idea
 
 GmailBundler/appsettings.user.json
+GmailBundler.App/appsettings.user.json
