+++ GmailBundler.App/appsettings.json
