--- GmailBundler.Csv.Interfaces/ICsvConverter.cs
@@ -1,8 +1,0 @@
-using GmailBundler.Dto;
-
-namespace GmailBundler.Csv.Interfaces;
-
-public interface ICsvConverter
-{
-    string Create(Gmail gmail);
-}
