--- GmailBundler.Csv/GmailBundler.Csv.csproj
@@ -1,19 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net7.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <PackageReference Include="Microsoft.Extensions.Options" Version="7.0.1" />
-        <PackageReference Include="RecordParser" Version="1.3.0" />
-        <PackageReference Include="Serilog" Version="3.0.0-dev-01993" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\GmailBundler.Configuration\GmailBundler.Configuration.csproj" />
-      <ProjectReference Include="..\GmailBundler.Csv.Interfaces\GmailBundler.Csv.Interfaces.csproj" />
-    </ItemGroup>
-</Project>
