+++ GmailBundler.Domain/Models/Gmail.cs
@@ -1,0 +1,3 @@
+﻿namespace GmailBundler.Domain.Models;
+
+public sealed record Gmail(string Id, string From, string Subject, string Label, DateTime Date);
