+++ GmailBundler.Domain/Models/GmailQuery.cs
@@ -1,3 +1,3 @@
-﻿namespace GmailBundler.Dto;
+﻿namespace GmailBundler.Domain.Models;
 
 public sealed record GmailQuery(string Query, string Label);
