--- GmailBundler.Downloader.Interfaces/GmailBundler.Downloader.Interfaces.csproj
@@ -1,13 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net7.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\GmailBundler.Dto\GmailBundler.Dto.csproj" />
-    </ItemGroup>
-
-</Project>
