--- GmailBundler.Downloader.Interfaces/IGmailDownloader.cs
@@ -1,9 +1,0 @@
-﻿using System.Runtime.CompilerServices;
-using GmailBundler.Dto;
-
-namespace GmailBundler.Downloader.Interfaces;
-
-public interface IGmailDownloader
-{
-    IAsyncEnumerable<Gmail> DownloadMails(GmailQuery query, [EnumeratorCancellation] CancellationToken cancellationToken);
-}
