--- GmailBundler.Downloader.Interfaces/IGmailServiceWrapper.cs
@@ -1,8 +1,0 @@
-﻿using GmailBundler.Dto;
-
-namespace GmailBundler.Downloader.Interfaces;
-
-public interface IGmailServiceWrapper
-{
-    Task Do(IEnumerable<GmailQuery> queries, CancellationToken cancellationToken);
-}
