--- GmailBundler.Dto/Gmail.cs
@@ -1,3 +1,0 @@
-﻿namespace GmailBundler.Dto;
-
-public sealed record Gmail(string Id, string From, string Subject, string Label, DateTime Date);
