--- GmailBundler/GmailBundler.csproj
@@ -1,44 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <OutputType>Exe</OutputType>
-        <TargetFramework>net7.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-        <TieredPGO>true</TieredPGO>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="7.0.0" />
-        <PackageReference Include="Microsoft.Extensions.Hosting" Version="7.0.1" />
-        <PackageReference Include="Microsoft.Extensions.Options" Version="7.0.1" />
-        <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="7.0.0" />
-        <PackageReference Include="Serilog" Version="3.0.0-dev-01993" />
-        <PackageReference Include="Serilog.AspNetCore" Version="7.0.0-dev-00302" />
-        <PackageReference Include="Serilog.Enrichers.Thread" Version="3.2.0-dev-00752" />
-        <PackageReference Include="Serilog.Extensions.Logging" Version="7.0.0-dev-10346" />
-        <PackageReference Include="Serilog.Extensions.Logging.File" Version="3.0.1-dev-00077" />
-        <PackageReference Include="Serilog.Settings.Configuration" Version="7.0.0-dev-00525" />
-        <PackageReference Include="Serilog.Sinks.Console" Version="4.1.1-dev-00910" />
-        <PackageReference Include="Serilog.Sinks.File" Version="5.0.1-dev-00947" />
-        <PackageReference Include="Google.Apis.Gmail.v1" Version="1.60.0.3035" />
-    </ItemGroup>
-
-    <ItemGroup>
-        <ProjectReference Include="..\GmailBundler.Configuration\GmailBundler.Configuration.csproj" />
-        <ProjectReference Include="..\GmailBundler.Csv.Interfaces\GmailBundler.Csv.Interfaces.csproj" />
-        <ProjectReference Include="..\GmailBundler.Csv\GmailBundler.Csv.csproj" />
-        <ProjectReference Include="..\GmailBundler.Downloader.Interfaces\GmailBundler.Downloader.Interfaces.csproj" />
-        <ProjectReference Include="..\GmailBundler.Downloader\GmailBundler.Downloader.csproj" />
-        <ProjectReference Include="..\GmailBundler.Dto\GmailBundler.Dto.csproj" />
-    </ItemGroup>
-
-    <ItemGroup>
-        <Content Include="appsettings.json">
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </Content>
-        <Content Include="appsettings.user.json">
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </Content>
-    </ItemGroup>
-</Project>
